LA-224: Add validation for credit commitments ignore and reasonToIgnore fields
