+++ src/test/java/com/selina/lending/api/validator/CCReasonToIgnoreValidationTest.java
@@ -17,18 +17,23 @@
 
 package com.selina.lending.api.validator;
 
-import com.selina.lending.internal.dto.creditcommitments.ApplicantCreditCommitmentsDto;
-import com.selina.lending.internal.dto.creditcommitments.UpdateCreditCommitmentsRequest;
-import org.junit.jupiter.api.BeforeAll;
-import org.junit.jupiter.api.Test;
+import static org.hamcrest.MatcherAssert.assertThat;
+import static org.hamcrest.Matchers.equalTo;
+
+import java.util.List;
 
 import javax.validation.Validation;
 import javax.validation.Validator;
 import javax.validation.ValidatorFactory;
-import java.util.List;
 
-import static org.hamcrest.MatcherAssert.assertThat;
-import static org.hamcrest.Matchers.equalTo;
+import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.Test;
+
+import com.selina.lending.internal.dto.DetailDto;
+import com.selina.lending.internal.dto.creditcommitments.ApplicantCreditCommitmentsDto;
+import com.selina.lending.internal.dto.creditcommitments.CreditCommitmentsDetailDto;
+import com.selina.lending.internal.dto.creditcommitments.SystemDto;
+import com.selina.lending.internal.dto.creditcommitments.UpdateCreditCommitmentsRequest;
 
 class CCReasonToIgnoreValidationTest {
     private static Validator validator;
@@ -43,11 +48,14 @@ class CCReasonToIgnoreValidationTest {
     @Test
     void shouldInvalidateWhenIgnoreButReasonIsAbsent() {
         //Given
-        var updateCCRequest = UpdateCreditCommitmentsRequest.builder()
-                .applicants(List.of(ApplicantCreditCommitmentsDto.builder()
-                        .ignore(true)
-                        .build()))
-                .build();
+        var updateCCRequest = UpdateCreditCommitmentsRequest.builder().applicants(
+                List.of(ApplicantCreditCommitmentsDto.builder()
+                        .creditCommitments(CreditCommitmentsDetailDto.builder()
+                                .system(SystemDto.builder()
+                                        .detail(List.of(DetailDto.builder().ignore(true).build()))
+                                        .build())
+                                .build())
+                        .build())).build();
 
         //When
         var violations = validator.validate(updateCCRequest);
@@ -56,19 +64,22 @@ class CCReasonToIgnoreValidationTest {
         assertThat(violations.size(), equalTo(1));
 
         var violation = violations.iterator().next();
-        assertThat(violation.getPropertyPath().toString(), equalTo("applicants[0].reasonToIgnore"));
+        assertThat(violation.getPropertyPath().toString(), equalTo("applicants[0].creditCommitments.system.detail[0].reasonToIgnore"));
         assertThat(violation.getMessage(), equalTo("This field is required"));
     }
 
     @Test
     void shouldInvalidateWhenIgnoreButReasonIsNotAcceptable() {
         //Given
-        var updateCCRequest = UpdateCreditCommitmentsRequest.builder()
-                .applicants(List.of(ApplicantCreditCommitmentsDto.builder()
-                        .ignore(true)
-                        .reasonToIgnore("bla-bla")
-                        .build()))
-                .build();
+        var updateCCRequest = UpdateCreditCommitmentsRequest.builder().applicants(
+                List.of(ApplicantCreditCommitmentsDto.builder()
+                        .creditCommitments(CreditCommitmentsDetailDto.builder()
+                                .system(SystemDto.builder().detail(List.of(DetailDto.builder()
+                                        .ignore(true)
+                                        .reasonToIgnore("bla-bla")
+                                        .build())).build())
+                                .build())
+                        .build())).build();
 
         //When
         var violations = validator.validate(updateCCRequest);
@@ -77,19 +88,22 @@ class CCReasonToIgnoreValidationTest {
         assertThat(violations.size(), equalTo(1));
 
         var violation = violations.iterator().next();
-        assertThat(violation.getPropertyPath().toString(), equalTo("applicants[0].reasonToIgnore"));
+        assertThat(violation.getPropertyPath().toString(), equalTo("applicants[0].creditCommitments.system.detail[0].reasonToIgnore"));
         assertThat(violation.getMessage(), equalTo("value is not valid"));
     }
 
     @Test
     void shouldBeValidWhenIgnoreAndReasonAcceptable() {
         //Given
-        var updateCCRequest = UpdateCreditCommitmentsRequest.builder()
-                .applicants(List.of(ApplicantCreditCommitmentsDto.builder()
-                        .ignore(true)
-                        .reasonToIgnore("Item is disputed - Experian to be updated")
-                        .build()))
-                .build();
+        var updateCCRequest = UpdateCreditCommitmentsRequest.builder().applicants(
+                List.of(ApplicantCreditCommitmentsDto.builder()
+                        .creditCommitments(CreditCommitmentsDetailDto.builder()
+                                .system(SystemDto.builder().detail(List.of(DetailDto.builder()
+                                        .ignore(true)
+                                        .reasonToIgnore("Item is disputed - Experian to be updated")
+                                        .build())).build())
+                                .build())
+                        .build())).build();
 
         //When
         var violations = validator.validate(updateCCRequest);
@@ -101,9 +115,8 @@ class CCReasonToIgnoreValidationTest {
     @Test
     void shouldBeValidWhenDoNotPassIgnoreAndReasonToIgnore() {
         //Given
-        var updateCCRequest = UpdateCreditCommitmentsRequest.builder()
-                .applicants(List.of(ApplicantCreditCommitmentsDto.builder().build()))
-                .build();
+        var updateCCRequest = UpdateCreditCommitmentsRequest.builder().applicants(
+                List.of(ApplicantCreditCommitmentsDto.builder().build())).build();
 
         //When
         var violations = validator.validate(updateCCRequest);
@@ -115,11 +128,14 @@ class CCReasonToIgnoreValidationTest {
     @Test
     void shouldBeValidWhenPassIgnoreAsFalse() {
         //Given
-        var updateCCRequest = UpdateCreditCommitmentsRequest.builder()
-                .applicants(List.of(ApplicantCreditCommitmentsDto.builder()
-                        .ignore(false)
-                        .build()))
-                .build();
+        var updateCCRequest = UpdateCreditCommitmentsRequest.builder().applicants(
+                List.of(ApplicantCreditCommitmentsDto.builder()
+                        .creditCommitments(CreditCommitmentsDetailDto.builder()
+                                .system(SystemDto.builder()
+                                        .detail(List.of(DetailDto.builder().ignore(false).build()))
+                                        .build())
+                                .build())
+                        .build())).build();
 
         //When
         var violations = validator.validate(updateCCRequest);
@@ -127,5 +143,4 @@ class CCReasonToIgnoreValidationTest {
         //Then
         assertThat(violations.size(), equalTo(0));
     }
-
 }
